+++ Assets/314 Arts/MarsFPSKit/Scripts/Player/Vitals/Kit_SimpleVitals.cs
@@ -229,7 +229,7 @@ namespace MarsFPSKit
                 if (pb.isFirstPersonActive)
                 {
                     //Update hud
-                    pb.main.hud.DisplayHealth(vrd.hitPoints);
+                    pb.main.hud.DisplayHealth((int)vrd.hitPoints);
                     pb.main.hud.DisplayHurtState(vrd.hitAlpha);
                 }
                 //Return hit reactions
